requirements.txt
python-dotenv==1.0.1
openai==1.12.0
moviepy==1.0.3
Pillow==10.2.0
requests==2.31.0
elevenlabs==0.2.24
google-generativeai
numpy
scipy 